modularize extractSetScores
- plan to move this up in hierarchy
- can apply finalSetScores to live scoreboard and MatchTiles
